feat: unbundled document settings
